refactor useEntries
