--- src/hooks/use-entries.ts
@@ -1,25 +1,0 @@
-import { getAxios } from "../get-axios";
-import { useCallback, useState } from "react";
-import { GetEntriesHandlerResponse } from "../../pages/api/user/entries";
-import { SpentProjectHours } from "../server/utils";
-
-export const useEntries = () => {
-    const [ isLoading, setIsLoading ] = useState(false);
-    const [ entries, setEntries ] = useState<SpentProjectHours[]>([]);
-    const load = useCallback((from: string, to: string, uid: string = '', projectId?: number) => {
-        setIsLoading(true)
-        getAxios().get<GetEntriesHandlerResponse>(`/user/entries?from=${ from }&to=${ to }&uid=${ uid }&project_id=${ projectId }`)
-            .then(({ data }) => {
-                setEntries(data.entries)
-            })
-            .finally(() => {
-                setIsLoading(false);
-            });
-    }, []);
-
-    return {
-        load,
-        isLoading,
-        entries,
-    }
-}
